LayoutTests/storage/indexeddb/mozilla/readyState-expected.txt

Test IndexedDB readyState property during different operations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBTransaction = window.IDBTransaction || window.webkitIDBTransaction;
PASS IDBTransaction == null is false
IDBRequest = window.IDBRequest || window.webkitIDBRequest;
PASS IDBRequest == null is false
indexedDB.open(name, description)
PASS request.readyState is IDBRequest.LOADING
openSuccess():
PASS request.readyState is IDBRequest.DONE
db = event.target.result
request = db.setVersion('1')
PASS request.readyState is IDBRequest.LOADING
setupObjectStore():
PASS request.readyState is IDBRequest.DONE
Deleted all object stores.
objectStore = db.createObjectStore('foo');
key = 10;
request = objectStore.add({}, key);
PASS request.readyState is IDBRequest.LOADING
getRecord():
PASS request.readyState is IDBRequest.DONE
PASS event.target.result is key
request = objectStore.get(key);
PASS request.readyState is IDBRequest.LOADING
finalCheck():
PASS request.readyState is IDBRequest.DONE
PASS event.target.result == null is false
PASS successfullyParsed is true

TEST COMPLETE
